new react routes and form
