+++ my-app-frontend/src/App.css
@@ -15,21 +15,24 @@ body {
   color: rgb(194, 202, 239);
 }
 
-.homeDiv {
-  background-image: url("https://images.unsplash.com/photo-1495954380655-01609180eda3?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80");
+.App {
+  background-image: url("city.jpeg");
   background-repeat: no-repeat;
-  /* height: 100%;
-  width: 100%; */
-  display: block;
   background-position: center;
-  padding: 100px
+  position: relative;
+  height: 1000px;
+  background-size: cover;
 }
 
 .welcome {
   font-size: 250px;
   color: rgb(255, 255, 255, 0.7);
   text-align: center;
-  margin: 100px;
+  /* margin: 100px; */
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
 }
 
 .header {
@@ -41,6 +44,6 @@ body {
 
 .headerDiv {
   outline: solid;
-  margin-bottom: 40px;
+  margin-bottom: 30px;
   background-color: rgba(32, 30, 27, 0.972);
 }
