+++ my-app-frontend/src/Header.js
@@ -5,7 +5,7 @@ function Header() {
     return (
         <div className="headerDiv">
             <h1 className="header">Travel.</h1>
-            <NavLink activeClassName="nav" to="/">Home</NavLink>  |  <NavLink activeClassName="nav" to="/Cities">Cities</NavLink>
+            <NavLink to="/">Home</NavLink>  |  <NavLink to="/cities">Cities</NavLink>
             <br></br>
             <br></br>
             <br></br>
