+++ my-app-frontend/src/Home.js
@@ -1,15 +1,7 @@
 import React from 'react';
-import CityItems from './CityItems';
 import ActivityItems from './ActivityItems';
 
-function Home({cities, activities}) {
-
-    const mapCities = cities.map((city) => {
-        return <CityItems
-        id={city.id}
-        name={city.name}
-        />
-    })
+function Home({activities}) {
 
     const mapActivities = activities.map((activity) => {
         return <ActivityItems
@@ -22,15 +14,9 @@ function Home({cities, activities}) {
     })
 
 
-    
-    console.log(mapActivities)
-
     return(
     <div className="homeDiv">
-        <section>
-            <ul>{mapCities}</ul>
-            <ul>{mapActivities}</ul>
-        </section>
+        <ul>{mapActivities}</ul>
         <h1 className="welcome">welcome</h1>
     </div>
     )
