+++ my-app-frontend/src/NewCityForm.js
@@ -1,0 +1,13 @@
+function NewCityForm() {
+    return (
+        <div>
+            <h2>New City</h2>
+            <form>
+                <input type="text" name="name" placeholder="city name"/>
+                <button type="submit">Add City</button>
+            </form>
+        </div>
+    )
+}
+
+export default NewCityForm;
